modified:   www/README.txt 	deleted:    www/home-plugins/99_sample-widget/index.php 	modified:   www/home-plugins/99_sample-widget/sample-widget.php 	deleted:    www/home-plugins/index.php 	modified:   www/index.php 	modified:   www/lib/favicon/favicon.php 	deleted:    www/lib/favicon/index.php 	modified:   www/lib/header.php 	modified:   www/lib/home.php 	modified:   www/lib/htmlheaders.php 	modified:   www/lib/htmlheaders/charset.php 	modified:   www/lib/htmlheaders/favicon.php 	deleted:    www/lib/htmlheaders/index.php 	modified:   www/lib/htmlheaders/theme.php 	deleted:    www/lib/index.php 	deleted:    www/lib/login/index.php 	modified:   www/lib/login/login-config.php 	modified:   www/lib/login/login-form.php 	modified:   www/lib/login/login.php 	deleted:    www/lib/menu/index.php 	modified:   www/lib/menu/menu.css 	modified:   www/lib/menu/menu.php 	modified:   www/lib/prevent-direct.php 	modified:   www/lib/prevent-index.php 	deleted:    www/lib/theme.css 	new file:   www/lib/themes/default.css 	modified:   www/login-plugins/10_check-cookies/check-cookies.php 	deleted:    www/login-plugins/10_check-cookies/index.php 	deleted:    www/login-plugins/99_sample-widget/index.php 	modified:   www/login-plugins/99_sample-widget/sample-widget.php 	deleted:    www/login-plugins/index.php 	modified:   www/router.php 	modified:   www/sample-menu-addon/index.php 	modified:   www/sample-menu-addon/menu-addon.php 	modified:   www/sample-page/index.php 	modified:   www/setup.sh

+++ www/setup.sh
@@ -1,53 +1,58 @@
 #!/bin/sh
 echo 'home-plugins'
 cd home-plugins
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../lib/prevent-index.php index.php
 
 echo '99_sample-widget'
 cd 99_sample-widget
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../../lib/prevent-index.php index.php
 echo -n '' > disabled
 
 echo 'lib'
 cd ../../lib
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s prevent-index.php index.php
 
 echo 'favicon'
 cd favicon
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../prevent-index.php index.php
 
 echo 'login'
 cd ../login
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../prevent-index.php index.php
 
 echo 'menu'
 cd ../menu
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../prevent-index.php index.php
 
 echo 'htmlheaders'
 cd ../htmlheaders
-rm index.php
+[ -e index.php ] && rm index.php
+ln -s ../prevent-index.php index.php
+
+echo 'themes'
+cd ../themes
+[ -e index.php ] && rm index.php
 ln -s ../prevent-index.php index.php
 
 echo 'login-plugins'
 cd ../../login-plugins
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../lib/prevent-index.php index.php
 
 echo '10_check-cookies'
 cd 10_check-cookies
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../../lib/prevent-index.php index.php
 
 echo '99_sample-widget'
 cd ../99_sample-widget
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../../lib/prevent-index.php index.php
 echo -n '' > disabled
